LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-column-axis-self-baseline-synthesized-001-expected.txt

Horizontal grid and verticalRL item
ÉÉ É ÉÉÉ É ÉÉ É
É É ÉÉ
Horizontal grid and verticalLR item
ÉÉ É ÉÉÉ É ÉÉ É
É É ÉÉ
VerticalLR grid and Horizontal item
ÉÉ É ÉÉÉ É ÉÉ É
É É ÉÉ
VerticalRL grid and Horizontal item
ÉÉ É ÉÉÉ É ÉÉ É
É É ÉÉ

FAIL .grid 1 assert_equals: 
<div class="grid alignItemsBaseline verticalRL">
  <div class="firstRowFirstColumn horizontalTB" data-offset-x="100" data-offset-y="0" data-expected-width="200" data-expected-height="100">ÉÉ É ÉÉÉ É ÉÉ É</div>
  <div class="firstRowSecondColumn bigFont paddingRight" data-offset-x="40" data-offset-y="100" data-expected-width="125" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 width25"></div>
</div>
offsetLeft expected 40 but got 175
FAIL .grid 2 assert_equals: 
<div class="grid alignItemsBaseline verticalLR">
  <div class="firstRowFirstColumn horizontalTB" data-offset-x="35" data-offset-y="0" data-expected-width="200" data-expected-height="100">ÉÉ É ÉÉÉ É ÉÉ É</div>
  <div class="firstRowSecondColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="100" data-expected-width="125" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 width25"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .grid 3 assert_equals: 
<div class="grid width300 alignItemsBaseline">
  <div class="firstRowFirstColumn verticalLR" data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="200">ÉÉ É ÉÉÉ É ÉÉ É</div>
  <div class="firstRowSecondColumn bigFont" data-offset-x="100" data-offset-y="160" data-expected-width="200" data-expected-height="100">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 height25"></div>
</div>
offsetTop expected 160 but got 0
FAIL .grid 4 assert_equals: 
<div class="grid width300 alignItemsBaseline">
  <div class="firstRowFirstColumn verticalRL" data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="200">ÉÉ É ÉÉÉ É ÉÉ É</div>
  <div class="firstRowSecondColumn bigFont" data-offset-x="100" data-offset-y="160" data-expected-width="200" data-expected-height="100">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 height25"></div>
</div>
offsetTop expected 160 but got 0
